limelight code
IT WORKS BABY WOOOOOO

+++ src/main/java/frc/robot/limelight.java
@@ -1,7 +1,9 @@
 package frc.robot;
 import frc.robot.LimelightHelpers;
+import java.text.DecimalFormat;
 
 public class limelight {
+
     public static double getX() {
         return LimelightHelpers.getTX("limelight");
 
@@ -16,9 +18,10 @@ public class limelight {
     }
     
     public static void target(){
+        
         getX();
         getY();
         getArea();
-        System.out.println("X: " + getX() + " Y: " + getY() + " Area: " + getArea());
+        System.out.println("X: " + (int)getX() + " Y: " + (int)getY() + " Area: " + (int)getArea());
     }
 }
